Make rplugin_Rgui global to avoid "invalid expression" error in windows.py
If rplugin_Rgui is local to the buffer, there is a bug created when multiple buffers are openened in R on windows. When you try to start an R terminal from any buffer other than the first that was opened, an error is thrown ("vim.error: invalid expression) from line 91 of windows.py, as the rplugin_Rgui variable is not defined.