Remove slidify support.
I am suggesting getting rid of slidify support because a) creating html
slides is easy enough with rmarkdown, b) development on the slidify
package seems to be slowing down, and c) adding support is easy enough
for the user to define their own RAction() call.

Also, this makes the code base cleaner with just a single vim function
responsible for all rmd markup.