Make more mappings respect vimrplugin_user_maps_only
There were two that I noticed:

RSendPartOfLine:

Some infrastructure was already set up for this (RNRightPart,
RIRightPart, RNLeftPart, RILeftPart) were already mentioned in the doc
but not really set up.  I also changed the function a little.  Sending
right would send the right part including the cursor and sending left
would send the left part without the cursor.  Now they both send the
cursor.

RNextRChunk/RPreviousChunk:

These were stealing gn and gN which are actually built-in vim commands.
The RCreateMaps puts LocalLeader in front of the maps so these are now
\gn and \gN.  I considered adding another flag in a:type to not include
LocalLeader in the maps, but I decided to keep it simple.  Instead I
was going to make a note in the doc about the change of these maps
(addition of LocalLeader) and how to map without the LocalLeader (just
put the original code into ftplugin/rmd.vim), but I wasn't sure where a
good place for this would be in the doc.